# docs/swagger.yaml
definitions:
...
  model.ConfigCondition:
    properties:
      config:
        $ref: '#/definitions/condition.Config'
      created_at:
        description: Created_at represent the creation date of the record.
        type: string
      id:
        description: ID is the unique identifier of the record in the database.
        type: string
      kind:
        description: Kind represent the kind of the resource configuration.
        type: string
      updated_at:
        description: Updated_at represent the last update date of the record.
        type: string
    type: object
  model.ConfigSource:
    properties:
      config:
        allOf:
        - $ref: '#/definitions/source.Config'
        description: Config is the configuration of the source resource.
      created_at:
        description: Created_at represent the creation date of the record.
        type: string
      id:
        description: ID is the unique identifier of the record in the database.
        type: string
      kind:
        description: Kind represent the kind of the resource configuration.
        type: string
      updated_at:
        description: Updated_at represent the last update date of the record.
        type: string
    type: object
...
  server.ConditionConfigResponse:
    properties:
      configs:
        description: Configs is a list of configuration conditions.
        items:
          $ref: '#/definitions/model.ConfigCondition'
        type: array
      total_count:
        description: TotalCount is the total number of conditions for pagination.
        type: integer
    type: object
  server.ConfigKindResponse:
    properties:
      data:
        items:
          type: string
        type: array
    type: object
  server.CreatePipelineReportResponse:
    properties:
      message:
        type: string
      reportid:
        type: string
    type: object
...
  server.DefaultResponseModel:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  server.GetPipelineReportByIDResponse:
    properties:
      data:
        $ref: '#/definitions/model.PipelineReport'
      latestReportByID:
        $ref: '#/definitions/model.PipelineReport'
      message:
        type: string
      nbReportsByID:
        type: integer
    type: object
  server.GetPipelineReportsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/database.SearchLatestReportData'
        type: array
      total_count:
        type: integer
    type: object
  server.ListLabelsResponse:
    properties:
      labels:
        description: Labels is a list of labels.
        items:
          $ref: '#/definitions/model.Label'
        type: array
      total_count:
        description: TotalCount is the total number of labels matching the query.
        type: integer
    type: object
  server.ListSCMsResponse:
    properties:
      scms:
        description: SCMs is a list of SCMs.
        items:
          $ref: '#/definitions/model.SCM'
        type: array
      total_count:
        description: TotalCount is the total number of SCMs matching the query.
        type: integer
    type: object
  server.SearchLabelsRequest:
    properties:
      end_time:
        description: |-
          EndTime is the end time for the time range filter
          This is optional and can be used to filter labels by a specific end time
          Time format is RFC3339: 2006-01-02T15:04:05Z07:00
        type: string
      id:
        description: Id is the unique identifier of the label.
        type: string
      key:
        description: Key is the key of the label.
        type: string
      key_only:
        description: KeyOnly specifies if we only need to retrieve a list of uniq
          Label keys
        type: boolean
      limit:
        description: |-
          Limit is the maximum number of labels to return
          This is optional and can be used to limit the number of labels returned
        type: integer
      page:
        description: |-
          Page is the page number for pagination
          This is optional and can be used to paginate the results
        type: integer
      start_time:
        description: |-
          StartTime is the start time for the time range filter
          This is optional and can be used to filter labels by a specific start time
          Time format is RFC3339: 2006-01-02T15:04:05Z07:00
        type: string
      value:
        description: Value is the value of the label.
        type: string
    type: object
  server.SearchPipelineReportsSummaryRequest:
    properties:
      days:
        description: |-
          Days is the number of days to summarize, today included.
          It defaults to 7 and is ignored when hours, or start_time and end_time, are provided.
        type: integer
      end_time:
        description: |-
          EndTime is the end time for the time range filter.
          Time format is: 2006-01-02 15:04:05Z07:00
        type: string
      granularity:
        description: |-
          Granularity is the size of the time buckets, one of "hour", "day", "week" or
          "month". It defaults to "day".
        type: string
      hours:
        description: |-
          Hours is the number of hours to summarize, the current hour included.
          It cannot be combined with days and is ignored when start_time and end_time are provided.
        type: integer
      labels:
        additionalProperties:
          type: string
        description: Labels is a map of labels to filter reports by.
        type: object
      metric:
        description: |-
          Metric is what the reports are counted by. It defaults to "result", which is
          the only value supported so far.
        type: string
      open_action:
        description: |-
          OpenAction filters reports by whether they carry an action left open, such as a
          pull request still waiting to be merged. This is optional: unset does not filter
          anything out, true only counts the reports with an open action and false only the
          ones without.

          The same breakdown is reported without filtering anything out under the open_actions
          key of every bucket.
        type: boolean
      results:
        description: |-
          Results is a list of pipeline results to filter reports by, such as
          "✔", "✗", "⚠" or "-". A report is counted when its result is any of them.
          An empty list does not filter anything out.
        items:
          type: string
        type: array
      scmid:
        description: |-
          ScmID is the ID of the SCM to filter reports by.
          Use "none" to only count the reports which are not attached to any SCM.
        type: string
      start_time:
        description: |-
          StartTime is the start time for the time range filter.
          Time format is: 2006-01-02 15:04:05Z07:00
        type: string
    type: object
  server.SearchPipelineReportsSummaryResponse:
    properties:
      data:
        description: Data contains one entry per time bucket, ordered from the oldest
          to the most recent one.
        items:
          $ref: '#/definitions/database.ReportResultSummaryEntry'
        type: array
      granularity:
        description: Granularity is the size of the time buckets of the entries.
        type: string
      metric:
        description: Metric is the metric the reports were counted by.
        type: string
      total_count:
        description: TotalCount is the total number of reports matching the query.
        type: integer
    type: object
...
  server.SourceConfigResponse:
    properties:
      configs:
        description: Configs is a list of configuration sources.
        items:
          $ref: '#/definitions/model.ConfigSource'
        type: array
      total_count:
        description: TotalCount is the total number of sources for pagination.
        type: integer
    type: object
  server.TargetConfigResponse:
    properties:
      configs:
        description: Configs is a list of configuration targets.
        items:
          $ref: '#/definitions/model.ConfigTarget'
        type: array
      total_count:
        description: TotalCount is the total number of targets for pagination.
        type: integer
    type: object
...
info:
  contact: {}
  description: API for managing Updatecli pipeline reports.
  title: Udash API
  version: "1.0"
paths:
...
  /api/pipeline/config/conditions:
    delete:
      description: Delete a configuration condition from the database
      parameters:
      - description: ID of the configuration condition to delete
        in: query
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
      summary: Delete a configuration condition
      tags:
      - Configuration Conditions
    get:
      description: List all configuration conditions from the database
      parameters:
      - description: ID of the configuration condition
        in: query
        name: id
        type: string
      - description: Kind of the configuration condition
        in: query
        name: kind
        type: string
      - description: Configuration of the condition
        in: query
        name: config
        type: string
      - description: Limit the number of reports returned, default is 100
        in: query
        name: limit
        type: string
      - description: Page number for pagination, default is 1
        in: query
        name: page
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.ConditionConfigResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
      summary: List all configuration conditions
      tags:
      - Configuration Conditions
  /api/pipeline/config/conditions/search:
    post:
      consumes:
      - application/json
      description: Search for configuration conditions in the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.ConditionConfigResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
      summary: Search configuration conditions
      tags:
      - Configuration Conditions
  /api/pipeline/config/kinds:
    post:
      consumes:
      - application/json
      description: Search for configuration by kind in the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.ConfigKindResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
      summary: Search configuration by kind
      tags:
      - Configuration
  /api/pipeline/config/sources:
    delete:
      description: Delete a configuration source from the database
      parameters:
      - description: ID of the configuration source to delete
        in: query
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
      summary: Delete a configuration source
      tags:
      - Configuration Sources
    get:
      description: List all configuration sources from the database
      parameters:
      - description: ID of the configuration source
        in: query
        name: id
        type: string
      - description: Kind of the configuration source
        in: query
        name: kind
        type: string
      - description: Configuration of the source
        in: query
        name: config
        type: string
      - description: Limit the number of reports returned, default is 100
        in: query
        name: limit
        type: string
      - description: Page number for pagination, default is 1
        in: query
        name: page
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.SourceConfigResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
      summary: List all configuration sources
      tags:
      - Configuration Sources
  /api/pipeline/config/sources/search:
    post:
      consumes:
      - application/json
      description: Search for configuration sources in the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.SourceConfigResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
      summary: Search configuration sources
      tags:
      - Configuration Sources
  /api/pipeline/config/targets:
    delete:
      description: Delete a configuration target from the database
      parameters:
      - description: ID of the configuration target to delete
        in: query
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
      summary: Delete a configuration target
      tags:
      - Configuration Targets
    get:
      description: List all configuration targets from the database
      parameters:
      - description: ID of the configuration target
        in: query
        name: id
        type: string
      - description: Kind of the configuration target
        in: query
        name: kind
        type: string
      - description: Configuration of the target
        in: query
        name: config
        type: string
      - description: Limit the number of reports returned, default is 100
        in: query
        name: limit
        type: string
      - description: Page number for pagination, default is 1
        in: query
        name: page
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.TargetConfigResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
      summary: List all target configurations.
      tags:
      - Configuration Targets
  /api/pipeline/config/targets/search:
    post:
      consumes:
      - application/json
      description: Search for configuration targets in the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.TargetConfigResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
      summary: Search configuration targets
      tags:
      - Configuration Targets
  /api/pipeline/labels:
    get:
      description: List labels data from the database with optional filtering
      parameters:
      - description: Filter by label ID
        in: query
        name: id
        type: string
      - description: Filter by label key
        in: query
        name: key
        type: string
      - description: Filter by label value
        in: query
        name: value
        type: string
      - description: Return only unique label keys (true/false)
        in: query
        name: keyonly
        type: string
      - description: Limit the number of labels returned, default is 100
        in: query
        name: limit
        type: string
      - description: Page number for pagination, default is 1
        in: query
        name: page
        type: string
      - description: Start time for filtering labels (RFC3339 format)
        in: query
        name: start_time
        type: string
      - description: End time for filtering labels (RFC3339 format)
        in: query
        name: end_time
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.ListLabelsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
      summary: List labels
      tags:
      - Labels
  /api/pipeline/labels/search:
    post:
      consumes:
      - application/json
      description: Search labels in the database using advanced filtering
      parameters:
      - description: Search parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server.SearchLabelsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.ListLabelsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
      summary: Search labels
      tags:
      - Labels
  /api/pipeline/reports:
    get:
      consumes:
      - application/json
      description: List all pipeline reports from the database
      parameters:
      - description: SCM ID
        in: query
        name: scmid
        type: string
      - description: Limit the number of reports returned, default is 100
        in: query
        name: limit
        type: string
      - description: Page number for pagination, default is 1
        in: query
        name: page
        type: string
      - description: Start time for filtering reports (RFC3339 format)
        in: query
        name: start_time
        type: string
      - description: End time for filtering reports (RFC3339 format)
        in: query
        name: end_time
        type: string
      - description: Only return the latest report per pipeline ID, default is true
        in: query
        name: latest
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.GetPipelineReportsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
      summary: List all pipeline reports
      tags:
      - Pipeline Reports
    post:
      consumes:
      - application/json
      description: Create a new pipeline report in the database
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/server.CreatePipelineReportResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
      summary: Create a new pipeline report
      tags:
      - Pipeline Reports
  /api/pipeline/reports/{id}:
    delete:
      description: Delete a pipeline report from the database
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
      summary: Delete a pipeline report
      tags:
      - Pipeline Reports
    get:
      description: Get the latest pipeline report for a specific ID
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.GetPipelineReportByIDResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
      summary: Get a pipeline report by ID
      tags:
      - Pipeline Reports
    put:
      consumes:
      - application/json
      description: Update a pipeline report in the database. Please note that this
        endpoint is not supported yet.
      parameters:
      - description: Report ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
      summary: Update a pipeline report
      tags:
      - Pipeline Reports
  /api/pipeline/reports/search:
    post:
      consumes:
      - application/json
      description: Search pipeline reports in the database using advanced filtering
      parameters:
      - description: Limit the number of reports returned, default is 100
        in: query
        name: limit
        type: string
      - description: Page number for pagination, default is 1
        in: query
        name: page
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.GetPipelineReportsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
      summary: Search pipeline reports
      tags:
      - Pipeline Reports
  /api/pipeline/reports/summary:
    post:
      consumes:
      - application/json
      description: |-
        Return the number of pipeline reports per result for each time bucket of the requested time range.
        Buckets are UTC hours, UTC calendar days, ISO weeks or calendar months depending on the
        granularity, and the date of an entry is the start of its bucket, formatted as RFC3339.
        Every report is counted, including several reports of the same pipeline, and buckets without
        any report are returned with a zeroed entry.
      parameters:
      - description: Summary filters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server.SearchPipelineReportsSummaryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.SearchPipelineReportsSummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
      summary: Summarize pipeline reports
      tags:
      - Pipeline Reports
  /api/pipeline/scms:
    get:
      description: List SCMs data from the database
      parameters:
      - description: ID of the SCM
        in: query
        name: scmid
        type: string
      - description: URL of the SCM
        in: query
        name: url
        type: string
      - description: Branch of the SCM
        in: query
        name: branch
        type: string
      - description: Return a summary of the SCMs
        in: query
        name: summary
        type: boolean
      - description: Limit the number of reports returned, default is 100
        in: query
        name: limit
        type: string
      - description: Page number for pagination, default is 1
        in: query
        name: page
        type: string
      - description: Start time for filtering SCMs (RFC3339 format)
        in: query
        name: start_time
        type: string
      - description: End time for filtering SCMs (RFC3339 format)
        in: query
        name: end_time
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.ListSCMsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
      summary: List SCMs
      tags:
      - SCMs
  /api/pipeline/scms/search:
    post:
      consumes:
      - application/json
      description: Search SCM data using JSON filters. When summary is true, the response
        contains SCM summary data for all matching SCMs.
      parameters:
      - description: SCM search filters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/server.SearchSCMsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.ListSCMsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.DefaultResponseModel'
      summary: Search SCMs
      tags:
      - SCMs
...
securityDefinitions:
  BearerAuth:
    description: Either an Udash API token, created from the tokens page and prefixed
      with "udash_pat_", or an access token from the configured identity provider.
      Send it as "Bearer <token>".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"